refactor: clean up argument parsing in consensus and meta scripts; update Dockerfile for version command
